requirements copy.txt
chardet==5.2.0
future==1.0.0
mysql-connector-python==9.0.0
pillow==10.4.0
reportlab==4.2.2
tkintertable==1.3.3